setup/requirements.txt
openai
pandas 
numpy 
flask 
flask_cors 
cassandra-driver 
tiktoken
tenacity
python-dotenv